no commit message given

+++ build.sh
@@ -1,5 +1,13 @@
 #!/bin/bash -e
-rm -rf toolchain
-mkdir toolchain
-sudo ./scripts/install_deps.sh
-scons toolchain
+if [ ! -f "deps.stamp" ]
+then
+	sudo ./scripts/install_deps.sh
+	touch "deps.stamp"
+fi
+if [ ! -f "toolchain.stamp" ]
+then
+	rm -rf toolchain
+	mkdir toolchain
+	scons toolchain
+	touch "toolchain.stamp"
+fi
